requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
altair==5.2.0
annotated-types==0.6.0
anyio==4.1.0
attrs==23.1.0
blinker==1.7.0
branca==0.6.0
cachetools==5.3.2
certifi==2023.7.22
charset-normalizer==3.3.0
click==8.1.7
contourpy==1.2.0
cramjam==2.7.0
cycler==0.12.1
distro==1.8.0
fastapi==0.109.2
fastparquet==2023.10.1
Flask==3.0.2
Flask-Testing==0.8.1
folium==0.14.0
fonttools==4.48.1
frozenlist==1.4.0
fsspec==2023.10.0
gitdb==4.0.11
GitPython==3.1.40
h11==0.14.0
httpcore==1.0.2
httpx==0.25.2
idna==3.4
importlib-metadata==6.11.0
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.20.0
jsonschema-specifications==2023.12.1
kiwisolver==1.4.5
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.8.2
mdurl==0.1.2
multidict==6.0.4
mysql-connector-python==8.3.0
numpy==1.26.1
openai==0.28.0
packaging==23.2
pandas==2.1.3
pillow==10.2.0
pluggy==1.4.0
protobuf==4.25.1
pyarrow==11.0.0
pydantic==2.5.2
pydantic_core==2.14.5
pydeck==0.8.1b0
Pygments==2.17.2
pyparsing==3.1.1
pytest==8.0.0
pytest-flask==1.3.0
python-dateutil==2.8.2
python-mnist==0.7
pytz==2023.3.post1
referencing==0.32.1
requests==2.31.0
requests-mock==1.11.0
rich==13.7.0
rpds-py==0.16.2
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
starlette==0.36.3
streamlit==1.29.0
tenacity==8.2.3
testcase==0.1.0
toml==0.10.2
toolz==0.12.0
tornado==6.4
tqdm==4.66.1
typing_extensions==4.9.0
tzdata==2023.3
tzlocal==5.2
urllib3==2.0.6
uvicorn==0.27.1
validators==0.22.0
Werkzeug==3.0.1
yarl==1.9.4
zipp==3.17.0